+++ lib/font.cpp
@@ -49,7 +49,7 @@ bool GlVisFont::LoadFont(const char* path, int font_size) {
     }
 
     //generate atlas
-    int w = 0, h = 0;
+    size_t w = 0, h = 0;
     for (int c = 32; c < 128; c++) {
         if (FT_Load_Char(face, c, FT_LOAD_RENDER)) {
             cout << "GLVis: Cannot load glyph: " << (char) c << endl;
